rename the SessionDetails layout into pascalcase naming style
