+++ client/src/Components/Layouts/SessionDetails/Content.js
