+++ client/src/Components/Layouts/SessionDetails/Header.js
