+++ client/src/Components/Layouts/SessionDetails/index.js
