+++ client/src/Components/Layouts/SessionDetails/styledComponents.js
